EnC test refactoring (#20061)
Refactor tests
