+++ src/EditorFeatures/CSharpTest/CSharpEditorServicesTest.csproj
@@ -206,6 +206,7 @@
     <Compile Include="Completion\CompletionProviders\ExplicitInterfaceTypeCompletionProviderTests.cs" />
     <Compile Include="Completion\CompletionProviders\ExplicitInterfaceMemberCompletionProviderTests.cs" />
     <Compile Include="AddUsing\AddUsingTestsWithAddImportDiagnosticProvider.cs" />
+    <Compile Include="EditAndContinue\StatementMatchingTests.cs" />
     <Compile Include="InitializeParameter\AddParameterCheckTests.cs" />
     <Compile Include="AddParameter\AddParameterTests.cs" />
     <Compile Include="AutomaticCompletion\AutomaticBraceCompletionTests.cs" />
@@ -359,10 +360,10 @@
     <Compile Include="EditAndContinue\CSharpEditAndContinueAnalyzerTests.cs" />
     <Compile Include="EditAndContinue\Helpers\CSharpEditAndContinueTestHelpers.cs" />
     <Compile Include="EditAndContinue\Helpers\Extensions.cs" />
-    <Compile Include="EditAndContinue\Helpers\RudeEditTestBase.cs" />
+    <Compile Include="EditAndContinue\Helpers\EditingTestBase.cs" />
     <Compile Include="EditAndContinue\LineEditTests.cs" />
-    <Compile Include="EditAndContinue\RudeEditStatementTests.cs" />
-    <Compile Include="EditAndContinue\RudeEditTopLevelTests.cs" />
+    <Compile Include="EditAndContinue\StatementEditingTests.cs" />
+    <Compile Include="EditAndContinue\TopLevelEditingTests.cs" />
     <Compile Include="EditAndContinue\SyntaxComparerTests.cs" />
     <Compile Include="EditAndContinue\SyntaxUtilitiesTests.cs" />
     <Compile Include="EncapsulateField\EncapsulateFieldCommandHandlerTests.cs" />
